Extracted common logic into abstract class on Root
