+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Container/CompilerPasses/RegisterAccessControlRuleBlockCompilerPass.php
@@ -18,8 +18,4 @@ class RegisterAccessControlRuleBlockCompilerPass extends AbstractInjectServiceDe
     {
         return AbstractAccessControlRuleBlock::class;
     }
-    protected function getRegistryMethodCallName(): string
-    {
-        return 'addServiceDefinitionID';
-    }
 }
