+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Registries/AccessControlRuleBlockRegistry.php
@@ -4,22 +4,8 @@ declare(strict_types=1);
 
 namespace GraphQLAPI\GraphQLAPI\Registries;
 
-class AccessControlRuleBlockRegistry implements AccessControlRuleBlockRegistryInterface
-{
-    /**
-     * @var string[]
-     */
-    protected array $serviceDefinitionIDs = [];
+use PoP\Root\Registries\AbstractServiceDefinitionIDRegistry;
 
-    public function addServiceDefinitionID(string $serviceDefinitionID): void
-    {
-        $this->serviceDefinitionIDs[] = $serviceDefinitionID;
-    }
-    /**
-     * @return string[]
-     */
-    public function getServiceDefinitionIDs(): array
-    {
-        return $this->serviceDefinitionIDs;
-    }
+class AccessControlRuleBlockRegistry extends AbstractServiceDefinitionIDRegistry implements AccessControlRuleBlockRegistryInterface
+{
 }
