+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Registries/AccessControlRuleBlockRegistryInterface.php
@@ -4,11 +4,8 @@ declare(strict_types=1);
 
 namespace GraphQLAPI\GraphQLAPI\Registries;
 
-interface AccessControlRuleBlockRegistryInterface
+use PoP\Root\Registries\ServiceDefinitionIDRegistryInterface;
+
+interface AccessControlRuleBlockRegistryInterface extends ServiceDefinitionIDRegistryInterface
 {
-    public function addServiceDefinitionID(string $serviceClass): void;
-    /**
-     * @return string[]
-     */
-    public function getServiceDefinitionIDs(): array;
 }
